+++ alicloud/resource_alicloud_ess_scalinggroup.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"github.com/denverdino/aliyungo/ess"
 	"github.com/hashicorp/terraform/helper/schema"
-	"strings"
 )
 
 func resourceAlicloudEssScalingGroup() *schema.Resource {
@@ -196,8 +195,8 @@ func buildAlicloudEssScalingGroupArgs(d *schema.ResourceData, meta interface{}) 
 
 	lbs, ok := d.GetOk("loadbalancer_ids")
 	if ok {
-		lbsStrings := lbs.([]interface{})
-		args.LoadBalancerIds = strings.Join(expandStringList(lbsStrings), COMMA_SEPARATED)
+		//lbsStrings := lbs.([]interface{})
+		args.LoadBalancerIds = convertListToJsonString(lbs.([]interface{}))
 	}
 
 	return args, nil